Posts

+++ src/main/java/nl/mengelmoestuintjes/gardening/service/UserService.java
@@ -1,10 +1,12 @@
 package nl.mengelmoestuintjes.gardening.service;
 
 import nl.mengelmoestuintjes.gardening.controller.exceptions.BadRequestException;
-import nl.mengelmoestuintjes.gardening.controller.exceptions.InvalidPasswordException;
+import nl.mengelmoestuintjes.gardening.controller.exceptions.InvalidException;
 import nl.mengelmoestuintjes.gardening.controller.exceptions.NotAuthorizedException;
 import nl.mengelmoestuintjes.gardening.controller.exceptions.UserNotFoundException;
 import nl.mengelmoestuintjes.gardening.dto.request.UserRequest;
+import nl.mengelmoestuintjes.gardening.dto.response.UserResponse;
+import nl.mengelmoestuintjes.gardening.model.posts.Post;
 import nl.mengelmoestuintjes.gardening.model.users.Authority;
 import nl.mengelmoestuintjes.gardening.model.users.Province;
 import nl.mengelmoestuintjes.gardening.model.users.User;
@@ -17,6 +19,7 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
+import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -35,115 +38,130 @@ public class UserService {
     public boolean userExists(String username) {
         return repository.existsById(username);
     }
+    private String getCurrentUserName() {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        return ((UserDetails) authentication.getPrincipal()).getUsername();
+    }
 
     // CREATE
     public String create(UserRequest request) {
         try {
-            String encryptedPassword = passwordEncoder.encode( request.getPassword());
-            User user = new User();
-            user.setUsername(request.getUsername());
-            user.setPassword(encryptedPassword);
-            user.setEnabled(true);
-            user.addAuthority("ROLE_USER");
-            user.setEmail(request.getEmail());
-            user.setLevel(request.getLevel());
-            user.setXp(request.getXp());
-            user.setLevelUpLimit(request.getLevelUpLimit());
-            user.setName(request.getName());
-            user.setBirthday(request.getBirthday());
-            user.setProvince(request.getProvince());
-            user.setMemberSince( LocalDate.now() );
-            user.setLastActivity();
-            User newUser = repository.save(user);
-            return newUser.getUsername();
+            request.setMemberSince(LocalDate.now());
+
+            UserResponse dto = new UserResponse();
+            User user = dto.toUser(request);
+
+            try {
+                String encryptedPassword = passwordEncoder.encode(request.getPassword());
+                user.setPassword(encryptedPassword);
+            } catch (Exception e) {
+                throw new InvalidException("password");
+            }
+
+            repository.save(user);
+            return user.getUsername();
         }
         catch (Exception ex) {
             throw new BadRequestException("Cannot create user.");
         }
     }
 
-
     // READ
-    private String getCurrentUserName() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        return ((UserDetails) authentication.getPrincipal()).getUsername();
+    public Iterable<User> getAll( String email, String level, Province province ) {
+        try {
+            boolean hasEmail = !email.isBlank();
+            boolean hasLevel = !level.isBlank();
+            boolean hasProvince = province != null;
+
+            if (hasEmail) return repository.findAllByEmailContaining(email);
+            if (hasLevel) return repository.findByLevel(level);
+            if (hasProvince) return repository.findAllByProvince(province);
+
+            return repository.findAll();
+        } catch (Exception e) {
+            throw new BadRequestException();
+        }
     }
-
-    public Iterable<User> getAll(
-            String email, String level, Province province
-    ) {
-        if ( !email.isBlank() ) return repository.findAllByEmailContaining(email);
-        if ( !level.isBlank() ) return repository.findByLevel( level );
-        if ( province != null ) return repository.findAllByProvince(province);
-
-        return repository.findAll();
-    }
-
     public User getUser(String username) {
         Optional<User> toFind = repository.findById( username );
         boolean userFound = toFind.isPresent();
-        if ( userFound ) {
-            return toFind.get();
-        } else {
-            throw new UserNotFoundException(username);
-        }
-    }
 
+        if ( userFound ) return toFind.get();
+        else throw new UserNotFoundException(username);
+    }
     public List<Authority> getAuthorities(String username) {
         User toFind = getUser( username );
         return toFind.getAuthorities();
     }
-
     public String getXP(String username) {
         User toFind = getUser( username );
         return toFind.getLevel() + ":" +toFind.getXp() + " -> " + toFind.getLevelUpLimit();
     }
+    public List<Post> getPosts(String username, boolean published) {
+        User toFind = getUser( username );
+        ArrayList<Post> posted = new ArrayList<>();
+        ArrayList<Post> concepts = new ArrayList<>();
 
+        for (Post p : toFind.getPosts()) {
+            if (p.isPublished()) {
+                posted.add(p);
+            } else {
+                concepts.add(p);
+            }
+        }
+
+        if (published) {
+            return posted;
+        } else {
+            return concepts;
+        }
+    }
+
+    // UPDATE
     public String addAuthority(String username, String authorityString) {
         User user = getUser( username );
         user.addAuthority(authorityString);
         repository.save(user);
-        return user.getUsername() + " has now authorithy " + authorityString;
+        return user.getUsername() + " has now authority " + authorityString;
     }
-
-    // UPDATE
-    public String changeCredentials(User toChange, User changed){
-        String out = toChange.getUsername() + " updated:";
+    public ArrayList<String> change(User toChange, UserRequest changed){
+        ArrayList<String> changedFields = new ArrayList<>();
 
         toChange.setPassword( passwordEncoder.encode(changed.getPassword()));
+        boolean enabled = toChange.isEnabled() != changed.isEnabled();
+        boolean email = !toChange.getEmail().equalsIgnoreCase(changed.getEmail());
+        boolean name = !toChange.getName().equalsIgnoreCase(changed.getName());
+        boolean province = toChange.getProvince() != changed.getProvince();
 
-        if (toChange.isEnabled() != changed.isEnabled()) {
+        if (enabled) {
             toChange.setEnabled(changed.isEnabled());
-            out += " enabled";
+            changedFields.add("enabled");
         }
-        if (!toChange.getEmail().equalsIgnoreCase(changed.getEmail())) {
+        if (email) {
             toChange.setEmail(changed.getEmail());
-            out += " email";
+            changedFields.add("email");
         }
-        if (!toChange.getName().equalsIgnoreCase(changed.getName())) {
+        if (name) {
             toChange.setName(changed.getName());
-            out += " name";
+            changedFields.add("name");
         }
-        if (toChange.getProvince() != changed.getProvince() ) {
+        if (province) {
             toChange.setProvince(changed.getProvince());
-            out += " province";
+            changedFields.add("province");
         }
 
-        if (out.equalsIgnoreCase(toChange.getUsername() + " updated:")) {
-            out = "nothing had been updated";
+        if (changedFields.isEmpty()) {
+            changedFields.add("NOTHING");
         }
-        return out;
+        return changedFields;
     }
-    public String update(String username, User modified) {
-        Optional<User> toFind = repository.findById(username);
-        if (toFind.isEmpty()) { throw new UserNotFoundException(username);
-        } else {
-            User user = toFind.get();
-            String changed = changeCredentials( user, modified );
-            repository.save(user);
-            return changed;
-        }
+    public ArrayList<String> update(String username, UserRequest modified) {
+        User user = getUser( username );
+        ArrayList<String> changed = change( user, modified );
+        repository.save(user);
+        return changed;
     }
+
     public void setPassword(String username, String password) {
         if (username.equals(getCurrentUserName())) {
             if (isValidPassword(password)) {
@@ -153,47 +171,53 @@ public class UserService {
                     user.setPassword(passwordEncoder.encode(password));
                     repository.save(user);
                 } else { throw new UserNotFoundException(username); }
-            } else { throw new InvalidPasswordException(); }
+            } else { throw new InvalidException("password"); }
         } else { throw new NotAuthorizedException(); }
     }
     public LocalDate setBirthday(String username, LocalDate newBirthday) {
-        User user = getUser( username );
-        user.setBirthday( newBirthday );
-        repository.save(user);
-        return newBirthday;
+        User user = getUser(username);
+        try {
+            user.setBirthday(newBirthday);
+            repository.save(user);
+            return newBirthday;
+        } catch (Exception e) {
+            throw new InvalidException("birthday");
+        }
     }
     public String setXP(String username, long toAdd) {
-        User user = getUser( username );
-        String xp = "" + toAdd;
-        String out = user.setXp( xp );
-        repository.save( user );
-        return out;
+        User user = getUser(username);
+        try {
+            String xp = "" + toAdd;
+            String out = user.setXp(xp);
+            repository.save(user);
+            return out;
+        } catch (Exception e) {
+            throw new InvalidException("set xp to user " + username);
+        }
     }
     public String setLastActivity(String username) {
         User user = getUser( username );
-        user.setLastActivity();
-        repository.save( user );
-        return "last activity = " + user.getLastActivity();
+        try {
+            user.setLastActivity();
+            repository.save(user);
+            return "last activity = " + user.getLastActivity();
+        } catch (Exception e) {
+            throw new InvalidException("last activity");
+        }
     }
 
     // DELETE
     public void deleteUser(String username) {
-        if (repository.existsById(username)) {
-            repository.deleteById(username);
-        }
-        else {
-            throw new UserNotFoundException(username);
-        }
+        if (userExists( username )) repository.deleteById(username);
+        else throw new UserNotFoundException(username);
     }
-
     public void removeAuthority(String username, String authorityString) {
-        Optional<User> userOptional = repository.findById(username);
-        if (userOptional.isEmpty()) {
-            throw new UserNotFoundException(username);
-        } else {
-            User user = userOptional.get();
+        User user = getUser( username );
+        try {
             user.removeAuthority(authorityString);
             repository.save(user);
+        } catch (Exception e) {
+            throw new InvalidException("cannot remove authority " + authorityString + " from user " + username);
         }
     }
 
@@ -219,21 +243,4 @@ public class UserService {
 
         return validPassword;
     }
-    public Long convertStringToLong(String s) {
-        if (s.equals("MAX")) {
-            return (long) 0;
-        }
-        return Long.parseLong( s );
-    }
-    public boolean hasReachedLevelUpLimit(long current, String limit) {
-        long levelUpLimit = convertStringToLong(limit);
-        return current >= levelUpLimit;
-    }
-    public boolean userIsMax(String level, String limit) {
-        boolean levelIsMax = level.equalsIgnoreCase("MAX");
-        boolean isLevel99 = level.equalsIgnoreCase("99");
-        boolean levelUpLimitMax = limit.equalsIgnoreCase("MAX");
-        return levelIsMax || isLevel99 || levelUpLimitMax;
-    }
-
 }
